update plugin media
