+++ src/media/registers/Routing.php
@@ -9,7 +9,7 @@ class Routing
     public static function registerEndpoints()
     {
         return [
-            'media' => [
+            'admin/media' => [
                 'fnc' => [
                     'get' => 'media.media.list',
                     'post' => 'media.media.list',
@@ -17,23 +17,16 @@ class Routing
                 'permission' => [
                     'get' => ['media_manager', 'media_read'],
                     'post' => ['media_manager', 'media_read'],
-                    'put' => ['media_manager', 'media_update'],
-                    'delete' => ['media_manager', 'media_delete']
                 ],
             ],
-            'media/upload' => [
+            'admin/media/upload' => [
                 'fnc' => [
                     'post' => 'media.media.upload',
                 ],
             ],
-            'media/ajax-upload' => [
+            'admin/media/ajax-upload' => [
                 'fnc' => [
                     'post' => 'media.ajax.upload',
-                ],
-            ],
-            'media/list' => [
-                'fnc' => [
-                    'post' => 'media.ajax.list',
                 ],
             ],
         ];
