+++ src/media/viewmodels/AdminMedia.php
@@ -11,104 +11,143 @@
 
 namespace DTM\media\viewmodels;
 
-use SPT\Web\ViewModel;
 use SPT\Web\Gui\Form;
+use SPT\Web\Gui\Listing;
+use SPT\Web\ViewModel;
 
-class AdminTag extends ViewModel
+class AdminMedia extends ViewModel
 {
     public static function register()
     {
         return [
-            'layout'=>'backend.tag.form',
-            'widget'=>[
-                'backend.javascript',
-                'backend.tags',
-            ],
+            'layout'=>[
+                'backend.list',
+                'backend.list.row',
+                'backend.list.filter'
+            ]
         ];
     }
-    
-    public function form()
+
+    public function list()
     {
-        $request = $this->container->get('request');
-        $UserEntity = $this->container->get('UserEntity');
-        $TagEntity = $this->container->get('TagEntity');
-        $router = $this->container->get('router');
+        $filter = $this->filter()['form'];
+        $limit  = $filter->getField('limit')->value;
+        $sort   = $filter->getField('sort')->value;
+        $search = trim($filter->getField('search')->value);
+        $page = $this->state('page', 1, 'int', 'get', 'media.page');
+        if ($page <= 0) $page = 1;
 
-        $urlVars = $request->get('urlVars');
-        $id = $urlVars ? (int) $urlVars['id'] : 0;
+        $where = [];
 
-        $data = $id ? $TagEntity->findByPK($id) : [];
+        if (!empty($search)) {
+            $where[] = "(`name` LIKE '%" . $search . "%') OR (`note` LIKE '%" . $search . "%')";
+        }
+
+        $start  = ($page - 1) * $limit;
+        $sort = $sort ? $sort : 'name asc';
+
+        $result = $this->MediaEntity->list($start, $limit, $where, $sort);
+        $total = $this->MediaEntity->getListTotal();
         
-        $form = new Form($this->getFormFields(), $data);
+        if (!$result) {
+            $result = [];
+            $total = 0;
+        }
 
+        $limit = $limit == 0 ? $total : $limit;
+        $list   = new Listing($result, $total, $limit, $this->getColumns());
         return [
-            'id' => $id,
-            'form' => $form,
-            'link_search' => $router->url('tag/search'),
-            'data' => $data,
+            'list' => $list,
+            'page' => $page,
+            'start' => $start,
+            'sort' => $sort,
+            'url' => $this->router->url(),
+            'link_list' => $this->router->url('admin/media'),
+            'title_page' => 'Media',
+            'link_form' => $this->router->url('admin/media/upload'),
+            'token' => $this->token->value(),
         ];
-        
     }
 
-    public function getFormFields()
+    public function getColumns()
     {
-        $fields = [
-            'name' => [
+        return [
+            'num' => '#',
+            'name' => 'Name',
+            'description' => 'Description',
+            'parent' => 'Parent',
+            'col_last' => ' ',
+        ];
+    }
+
+    protected $_filter;
+    public function filter()
+    {
+        if (null === $this->_filter) :
+            $data = [
+                'search' => $this->state('search', '', '', 'post', 'media.search'),
+                'limit' => $this->state('limit', 20, 'int', 'post', 'media.limit'),
+                'sort' => $this->state('sort', '', '', 'post', 'media.sort')
+            ];
+            $filter = new Form($this->getFilterFields(), $data);
+
+            $this->_filter = $filter;
+        endif;
+
+        return ['form' => $this->_filter];
+    }
+
+    public function getFilterFields()
+    {
+        return [
+            'search' => [
                 'text',
+                'default' => '',
                 'showLabel' => false,
-                'placeholder' => 'Tag Name',
-                'formClass' => 'form-control mb-3',
-                'required' => 'required',
+                'formClass' => 'form-control h-full input_common w_full_475',
+                'placeholder' => 'Search..'
             ],
-            'description' => [
-                'textarea',
-                'showLabel' => false,
-                'placeholder' => 'Description',
-                'formClass' => 'form-control',
-            ],
-            'parent_id' => [
+            'status' => [
                 'option',
-                'type' => 'select2',
+                'default' => '1',
                 'formClass' => 'form-select',
-                'options' => [],
-                'showLabel' => false,
-                'placeholder' => 'Tags',
-                'formClass' => 'form-control',
+                'options' => [
+                    ['text' => '--', 'value' => ''],
+                    ['text' => 'Show', 'value' => '1'],
+                    ['text' => 'Hide', 'value' => '0'],
+                ],
+                'showLabel' => false
             ],
-            'token' => ['hidden',
-                'default' => $this->container->get('token')->value(),
+            'limit' => [
+                'option',
+                'formClass' => 'form-select',
+                'default' => 20,
+                'options' => [
+                    ['text' => '20', 'value' => 20],
+                    ['text' => '50', 'value' => 50],
+                    ['text' => 'All', 'value' => 0],
+                ],
+                'showLabel' => false
             ],
+            'sort' => [
+                'option',
+                'formClass' => 'form-select',
+                'default' => 'name asc',
+                'options' => [
+                    ['text' => 'name ascending', 'value' => 'name asc'],
+                    ['text' => 'name descending', 'value' => 'name desc'],
+                ],
+                'showLabel' => false
+            ]
         ];
-
-        return $fields;
     }
 
-    public function javascript()
+    public function row($layoutData, $viewData)
     {
+        $row = $viewData['list']->getRow();
         return [
-            'link_tag' => $this->router->url('tag/search'),
-            'link_add_tag' => $this->router->url('tag/ajax-add'),
-        ];
-    }
-
-    public function tags($layoutData, $viewData)
-    {
-        $data = isset($viewData['data']) ? $viewData['data'] : [];
-        $tags = isset($data['tags']) ? $data['tags'] : '';
-
-        $tags = $this->TagModel->convert($tags, false);
-        if ($tags)
-        {
-            $where = ['id IN ('. implode(',', $tags) .')'];
-            $tags = $this->TagEntity->list(0, 0, $where);
-        }
-        else
-        {
-            $tags = [];
-        }
-        
-        return [
-            'tags' => $tags,
+            'item' => $row,
+            'index' => $viewData['list']->getIndex(),
         ];
     }
 }
