+++ src/media/views/layouts/backend/form.php
@@ -1,0 +1,27 @@
+<div class="modal fade" id="formEditTag" aria-labelledby="formEditTagLabel" tabindex="-1" aria-hidden="true" style="display: none;">
+    <div class="modal-dialog modal-dialog-centered " style="max-width: 600px;">
+        <div class="modal-content container px-5 pt-5">
+            <form action="<?php echo $this->link_form ?>" enctype="multipart/form-data"  method="post" id="form_tag">
+                <div class="row g-3 align-items-center">
+                    <div class="row px-0">
+                        <input type="file" class="form-control" required multiple name="file[]">
+                    </div>
+                    <div class="row g-3 align-items-center m-0">
+                        <div class="modal-footer">
+                            <?php $this->ui->field('token'); ?>
+                            <input class="form-control rounded-0 border border-1" id="_method" type="hidden" name="_method" value="POST">
+                            <div class="row">
+                                <div class="col-6 text-end pe-0">
+                                    <button type="button" class="btn btn-outline-secondary fs-4" data-bs-dismiss="modal">Cancel</button>
+                                </div>
+                                <div class="col-6 text-end pe-0 ">
+                                    <button type="submit" class="btn btn-outline-success fs-4">Upload</button>
+                                </div>
+                            </div>
+                        </div>
+                    </div>
+                </div>
+            </form>
+        </div>
+    </div>
+</div>
